Create iris user and group in iris-web role.
In the past, we create user/group in iris.spec while the rpm
package is installed. It's better to maintain the logic here to
keep all deployment things together.

Add there default varaibles:
- sys_user
- sys_group
- sys_user_home

Since there are default variables, if we don't need to modify
then we can omit them in host_vars files.

Change-Id: I8127f5e359a4602691c48be41a4e7cf2112ab911