Install iris package by pip.
It will also install its requirements automatically.

Current we are using sdist packaging format. It can be a local
tar file, a remote url or a package name in pypi.

For testing we will use tar file or url in internal file server,
for final release we will use pypi.

Change-Id: Icceeede5db4c7bb3fc0e686c6e3ea88fa3de6194